+++ pyxel/util/__init__.py
@@ -21,6 +21,7 @@ from .timing import time_pipeline
 from .add_model import create_model
 from .randomize import set_random_seed
 from .image import fit_into_array, load_cropped_and_aligned_image
+from .caching import get_cache
 
 __all__ = [
     "convert_to_int",